Keep trying to bind to HTTP port
This patch puts the HTTP server's ListenAndConnect into an infinte loop
so that if there is an error connecting we will keep trying until it is
successful. Right now the timeout for errors is 5 seconds.
The goal of this patch is to prevent a production issue from happening
again...the batchs server was restarted at one point and for some reason
it couldn't immediately bind to port 15000, and so it never started
listening on the port and other services were then erroring out. Yet the
batch processing was working just fine.