Put filename-normalize errors in the LOG file
The err file code seems like it is working around rof not returning
correct status codes (0 = ok, >0 for error). The rof tool has been
updated to return error status codes in this situation.

This will make sure the error messages from the ROF tool end up in the
LOG file so all the errors are in one place.