Pipeline fixes
This moves from using a CodeBuild to using a CodePipeline to allow for more easily adding future stages to this process such as deploying to a test env, running tests, approving changes, etc. Unfortunately, CodePipeline does not give you the .git directory in its artifacts and this project requires it since it uses git-describe to create the rpm's filename based on the version of the repo. So, before running the make in the buildspec, I had to checkout another copy of the code into the target /go path using the same oauth token that creates the github webhook.